+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes for Crate Admin Interface
 Unreleased
 ==========
 
+- Fixed the table/column display to no longer display all columns as being
+  generated columns.
+
 - Replaced the Slack icon with Discourse in the help section, updated nametags
   for all languages and translations for Spanish.